refactor(test classes): Изменено расположение 2 тестовых классов
Тесты TestBuilderBankAccount, TestBuyer и TestFactory перенесены в соответствующие директории.
Отредактирован файл README.md.
